refactor: rename login register to dialog
