+++ frontend/src/App.tsx
@@ -3,7 +3,7 @@ import React from "react";
 import { Provider } from "react-redux";
 import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
 import { APIProvider } from "./context/APIContext";
-import { LoginRegisterProvider } from "./context/LoginRegisterContext";
+import { DialogProvider } from "./context/DiagramContext";
 import { PurchasedOrderProvider } from "./context/PurchasedOrderContext";
 import AttractionPage from "./pages/AttractionPage";
 import BookingPage from "./pages/BookingPage";
@@ -19,7 +19,7 @@ function App() {
     <Provider store={store}>
       <APIProvider>
         {/* <APIProvider isMock> */}
-        <LoginRegisterProvider>
+        <DialogProvider>
           <Global styles={GlobalStyles} />
           <PurchasedOrderProvider>
             <Router>
@@ -37,7 +37,7 @@ function App() {
               </Routes>
             </Router>
           </PurchasedOrderProvider>
-        </LoginRegisterProvider>
+        </DialogProvider>
       </APIProvider>
     </Provider>
   );
